Bug 875062: Add URL shortening
