+++ kuma/core/helpers.py
@@ -3,6 +3,8 @@ import HTMLParser
 import os
 import urllib
 import urlparse
+import hashlib
+import bitly_api
 
 from django.http import QueryDict
 from django.utils.encoding import smart_str
@@ -86,6 +88,25 @@ def urlparams(url_, hash=None, query_dict=None, **query):
     new = urlparse.ParseResult(url_.scheme, url_.netloc, url_.path,
                                url_.params, query_string, fragment)
     return new.geturl()
+
+
+bitly = bitly_api.Connection(login=getattr(settings, 'BITLY_USERNAME', ''),
+                             api_key=getattr(settings, 'BITLY_API_KEY', ''))
+
+@register.filter
+def bitly_shorten(url):
+    """Attempt to shorten a given URL through bit.ly / mzl.la"""
+    cache_key = 'bitly:%s' % hashlib.md5(smart_str(url)).hexdigest()
+    short_url = memcache.get(cache_key)
+    if short_url is None:
+        try:
+            short_url = bitly.shorten(url)['url']
+            memcache.set(cache_key, short_url, 60 * 60 * 24 * 30 * 12)
+        except (bitly_api.BitlyError, KeyError):
+            # Just in case the bit.ly service fails or the API key isn't
+            # configured, fall back to using the original URL.
+            return url
+    return short_url
 
 
 class Paginator(object):
