+++ kuma/wiki/templates/wiki/document.html
@@ -337,12 +337,28 @@
               </article>
 
             {% if waffle.flag('share_links') %}
+                {% set full_url = request.build_absolute_uri(document.get_absolute_url()) %}
+
+                {% set share_url_twitter = full_url|urlparams(utm_campaign='share', utm_source='twitter.com', utm_medium='doc widget') %}
+                {% set share_url_facebook = full_url|urlparams(utm_campaign='share', utm_source='facebook.com', utm_medium='doc widget')%}
+                {% set share_url_gplus = full_url|urlparams(utm_campaign='share', utm_source='plus.google.com', utm_medium='doc widget')%}
+
+                {% set short_share_url_twitter = share_url_twitter | bitly_shorten %}
+                {% set short_share_url_facebook = share_url_facebook | bitly_shorten %}
+                {% set short_share_url_gplus = share_url_gplus | bitly_shorten %}
+
+                {% set share_text = 'I found this article about '+ document.title +' on MDN' %}
+
+                {% set twitter_url = 'http://twitter.com/share'|urlparams(url=short_share_url_twitter, text=share_text) %}
+                {% set facebook_url = 'http://facebook.com/sharer.php?s=100&p[url]='+short_share_url_facebook+'&p[images][0]=/media/redesign/img/favicon144.png&p[title]=' + share_text %}
+                {% set gplus_url = 'https://plus.google.com/share'|urlparams(url=short_share_url_gplus) %}
+
                 <div class="share">
                     <strong><i class="icon-share" aria-hidden="true"></i> {{ _('Share:') }}</strong>
                     <ul>
-                        <li><button class="share-link twitter"><i aria-hidden="true" class="icon-twitter"></i>{{ _('Twitter') }}</button></li>
-                        <li><button class="share-link facebook"><i aria-hidden="true" class="icon-facebook"></i>{{ _('Facebook') }}</button></li>
-                        <li><button class="share-link gplus"><i aria-hidden="true" class="icon-google-plus"></i>{{ _('Google +') }}</button></li>
+                        <li><a href="{{ twitter_url }}" target="_blank" data-url="{{short_share_url_twitter}}" data-title="{{ share_text }}" class="share-link twitter"><i aria-hidden="true" class="icon-twitter"></i>{{ _('Twitter') }}</a></li>
+                        <li><a href="{{ facebook_url }}" target="_blank" data-url="{{short_share_url_facebook}}" data-title="{{ share_text }}" class="share-link facebook"><i aria-hidden="true" class="icon-facebook"></i>{{ _('Facebook') }}</a></li>
+                        <li><a href="{{ gplus_url }}" target="_blank" data-url="{{short_share_url_gplus}}" data-title="{{ share_text }}" class="share-link gplus"><i aria-hidden="true" class="icon-google-plus"></i>{{ _('Google +') }}</a></li>
                     </ul>
                 </div>
             {% endif %}
